Refines domain entities and enhances repository capabilities
Ensures consistent primary key definitions for analytic entities using the `[Key]` attribute. Updates data types for entity properties to improve type safety and align with database mapping.

Introduces new message structures and queue settings to support aggregation worker processes via RabbitMQ. Enhances the base repository with explicit transaction management and `SaveChangesAsync` for improved data integrity and control over persistence operations.

+++ src/Core/Metriflow.Domain/Entities/TimeIntervalAnalytic.cs
@@ -7,6 +7,6 @@ public class TimeIntervalAnalytic : AggregateAnalytics
     public Guid Id { get; set; }
     public DateTime Date { get; set; }
     [ForeignKey("TimeInterval")]
-     public byte TimeIntervalId { get; set; }
+     public int TimeIntervalId { get; set; }
      public TimeInterval TimeInterval { get; set; }
 }
